refactor: `ua` locales -> `uk` [WTEL-5377](https://webitel.atlassian.net/browse/WTEL-5377)
