+++ src/app/locale/i18n.js
@@ -3,12 +3,12 @@ import { createI18n } from 'vue-i18n';
 import en from './en/en';
 import kz from './kz/kz';
 import ru from './ru/ru';
-import ua from './ua/ua';
+import uk from './uk/uk';
 
 const messages = {
   en,
   ru,
-  ua,
+  uk,
   kz,
 };
