+++ src/app/locale/uk/uk.js
