+++ src/app/plugins/webitel-ui.js
@@ -4,8 +4,7 @@ import WebitelUI from '@webitel/ui-sdk/dist/ui-sdk.js';
 import WebitelUIEn from '@webitel/ui-sdk/src/locale/en/en';
 import WebitelUIKz from '@webitel/ui-sdk/src/locale/kz/kz';
 import WebitelUIRu from '@webitel/ui-sdk/src/locale/ru/ru';
-import WebitelUIUa from '@webitel/ui-sdk/src/locale/ua/ua';
-import eventBus from '@webitel/ui-sdk/src/scripts/eventBus';
+import WebitelUIUk from '@webitel/ui-sdk/src/locale/uk/uk';
 
 import i18n from '../locale/i18n';
 
@@ -17,5 +16,5 @@ export default [WebitelUI, { eventBus, globals }];
 
 i18n.global.mergeLocaleMessage('en', WebitelUIEn);
 i18n.global.mergeLocaleMessage('ru', WebitelUIRu);
-i18n.global.mergeLocaleMessage('ua', WebitelUIUa);
+i18n.global.mergeLocaleMessage('uk', WebitelUIUk);
 i18n.global.mergeLocaleMessage('kz', WebitelUIKz);
